feat: 更新监控模版config
