--- src/datas/monitor-template.ts
@@ -1,183 +1,0 @@
-export const monitorJson = {
-  theme: 'light',
-  // layouts 部分，可以加上hash #mode=edit 然后页面拖拽调整完大小，替换保存就行(只需要 i、x、y、w、h)
-  layouts: {
-    root: {
-      lg: [
-        { w: 6, h: 5, x: 0, y: 0, i: 'a', moved: false, static: false },
-        { w: 6, h: 5, x: 6, y: 0, i: 'b', moved: false, static: false },
-        { w: 6, h: 5, x: 12, y: 0, i: 'c', moved: false, static: false },
-        { w: 6, h: 5, x: 18, y: 0, i: 'd', moved: false, static: false },
-        { w: 6, h: 11, x: 0, y: 5, i: "e", moved: false, static: false },
-        { w: 24, h: 11, x: 0, y: 16, i: "f", moved: false, static: false },
-        { w: 24, h: 12, x: 0, y: 28, i: "g", moved: false, static: false }
-      ],
-      xs: [
-        { i: ' a', x: 0, y: 0, w: 12, h: 6 },
-        { i: ' b', x: 0, y: 6, w: 12, h: 8 },
-        { i: ' c', x: 7, y: 22, w: 5, h: 13 },
-        { i: ' d', x: 0, y: 22, w: 7, h: 13 },
-        { i: ' e', x: 0, y: 14, w: 6, h: 8 },
-        { i: ' f', x: 6, y: 14, w: 6, h: 8 },
-        { i: ' g', x: 6, y: 35, w: 6, h: 14 },
-        { i: ' h', x: 0, y: 73, w: 6, h: 14 },
-        { i: ' i', x: 0, y: 49, w: 12, h: 13 },
-        { i: ' j', x: 0, y: 62, w: 12, h: 11 },
-        { i: ' k', x: 6, y: 73, w: 6, h: 14 },
-        { i: ' l', x: 0, y: 35, w: 6, h: 14 }]
-    },
-  },
-  content: {
-    tag: 'panel',
-    id: 'root',
-    attributes: {},
-    children: [
-      {
-        tag: 'statistic',
-        id: 'a',
-        attributes: {
-          measures: ['monitor_count'],
-          meta: {
-            monitor_count: {
-              alias: '监控总数',
-              icon: '<svg viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg"><path d="M689.00486 572.124909c79.006322-55.180406 130.7762-146.675053 130.7762-250.346961 0-168.551225-136.490232-305.173608-305.173608-305.173608-168.551225 0-305.087201 136.621367-305.087201 305.173608 0 103.580419 51.547254 194.921567 130.468185 250.079609C162.135773 641.653928 36.012326 814.3232 36.012326 1016.886178c0 0.086407 0 0.086407 0 0.086407l85.691159 0c0-0.217542 0-0.484895 0-0.707519 0-200.834842 151.237333-365.976556 346.051145-388.827599 1.990406-0.131135 54.644684-4.739159 97.912132 0.729883 1.065345 0.244989 2.217097 0.331396 3.324121 0.553004l0.484895 0c189.942501 27.190699 335.996442 190.119381 335.996442 387.544712 0 0.222625 0 0.489977 0 0.707519l87.023857 0c0 0 0 0 0-0.086407C992.496077 814.544808 866.635916 641.961943 689.00486 572.124909L689.00486 572.124909zM514.785349 550.557769c-126.259665 0-228.693415-102.34226-228.693415-228.64767S388.525684 93.262429 514.785349 93.262429c126.21392 0 228.64767 102.34226 228.64767 228.64767S640.999269 550.557769 514.785349 550.557769L514.785349 550.557769zM514.785349 550.557769" p-id="2751"></path></svg>'
-            },
-          },
-          data: [
-            {
-              monitor_count: 21,
-            },
-          ],
-        },
-      },
-      {
-        tag: 'statistic',
-        id: 'b',
-        attributes: {
-          measures: ['healthy_app_count'],
-          meta: {
-            healthy_app_count: {
-              alias: '健康应用数',
-              icon: '<svg viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg"><path d="M689.00486 572.124909c79.006322-55.180406 130.7762-146.675053 130.7762-250.346961 0-168.551225-136.490232-305.173608-305.173608-305.173608-168.551225 0-305.087201 136.621367-305.087201 305.173608 0 103.580419 51.547254 194.921567 130.468185 250.079609C162.135773 641.653928 36.012326 814.3232 36.012326 1016.886178c0 0.086407 0 0.086407 0 0.086407l85.691159 0c0-0.217542 0-0.484895 0-0.707519 0-200.834842 151.237333-365.976556 346.051145-388.827599 1.990406-0.131135 54.644684-4.739159 97.912132 0.729883 1.065345 0.244989 2.217097 0.331396 3.324121 0.553004l0.484895 0c189.942501 27.190699 335.996442 190.119381 335.996442 387.544712 0 0.222625 0 0.489977 0 0.707519l87.023857 0c0 0 0 0 0-0.086407C992.496077 814.544808 866.635916 641.961943 689.00486 572.124909L689.00486 572.124909zM514.785349 550.557769c-126.259665 0-228.693415-102.34226-228.693415-228.64767S388.525684 93.262429 514.785349 93.262429c126.21392 0 228.64767 102.34226 228.64767 228.64767S640.999269 550.557769 514.785349 550.557769L514.785349 550.557769zM514.785349 550.557769" p-id="2751"></path></svg>'
-            },
-          },
-          data: [
-            {
-              healthy_app_count: 19,
-            },
-          ],
-        },
-      },
-      {
-        tag: 'statistic',
-        id: 'c',
-        attributes: {
-          measures: ['anomaly_app_counts'],
-          meta: {
-            anomaly_app_counts: {
-              alias: '异常应用数',
-              icon: '<svg viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg"><path d="M689.00486 572.124909c79.006322-55.180406 130.7762-146.675053 130.7762-250.346961 0-168.551225-136.490232-305.173608-305.173608-305.173608-168.551225 0-305.087201 136.621367-305.087201 305.173608 0 103.580419 51.547254 194.921567 130.468185 250.079609C162.135773 641.653928 36.012326 814.3232 36.012326 1016.886178c0 0.086407 0 0.086407 0 0.086407l85.691159 0c0-0.217542 0-0.484895 0-0.707519 0-200.834842 151.237333-365.976556 346.051145-388.827599 1.990406-0.131135 54.644684-4.739159 97.912132 0.729883 1.065345 0.244989 2.217097 0.331396 3.324121 0.553004l0.484895 0c189.942501 27.190699 335.996442 190.119381 335.996442 387.544712 0 0.222625 0 0.489977 0 0.707519l87.023857 0c0 0 0 0 0-0.086407C992.496077 814.544808 866.635916 641.961943 689.00486 572.124909L689.00486 572.124909zM514.785349 550.557769c-126.259665 0-228.693415-102.34226-228.693415-228.64767S388.525684 93.262429 514.785349 93.262429c126.21392 0 228.64767 102.34226 228.64767 228.64767S640.999269 550.557769 514.785349 550.557769L514.785349 550.557769zM514.785349 550.557769" p-id="2751"></path></svg>'
-            },
-          },
-          data: [
-            {
-              anomaly_app_counts: 2,
-            },
-          ],
-        },
-      },
-      {
-        tag: 'statistic',
-        id: 'd',
-        attributes: {
-          measures: ['custom_count'],
-          meta: {
-            custom_count: {
-              alias: '自定义大盘',
-              icon: '<svg viewBox="0 0 1024 1024" version="1.1" xmlns="http://www.w3.org/2000/svg"><path d="M689.00486 572.124909c79.006322-55.180406 130.7762-146.675053 130.7762-250.346961 0-168.551225-136.490232-305.173608-305.173608-305.173608-168.551225 0-305.087201 136.621367-305.087201 305.173608 0 103.580419 51.547254 194.921567 130.468185 250.079609C162.135773 641.653928 36.012326 814.3232 36.012326 1016.886178c0 0.086407 0 0.086407 0 0.086407l85.691159 0c0-0.217542 0-0.484895 0-0.707519 0-200.834842 151.237333-365.976556 346.051145-388.827599 1.990406-0.131135 54.644684-4.739159 97.912132 0.729883 1.065345 0.244989 2.217097 0.331396 3.324121 0.553004l0.484895 0c189.942501 27.190699 335.996442 190.119381 335.996442 387.544712 0 0.222625 0 0.489977 0 0.707519l87.023857 0c0 0 0 0 0-0.086407C992.496077 814.544808 866.635916 641.961943 689.00486 572.124909L689.00486 572.124909zM514.785349 550.557769c-126.259665 0-228.693415-102.34226-228.693415-228.64767S388.525684 93.262429 514.785349 93.262429c126.21392 0 228.64767 102.34226 228.64767 228.64767S640.999269 550.557769 514.785349 550.557769L514.785349 550.557769zM514.785349 550.557769" p-id="2751"></path></svg>'
-            },
-          },
-          data: [
-            {
-              custom_count: 5,
-            },
-          ],
-        },
-      },
-      {
-        tag: 'liguid',
-        id: 'e',
-        attributes: {
-          title: '访问频次发布',
-          width: 230,
-          height: 230,
-          autoFit: true,
-          percent: 0.32,
-          outline: {
-            distance: 2,
-          },
-          statistic: {
-            title: {
-              formatter: () => '资源剩余',
-            },
-            content: {
-              customHtml: (container, view, { percent }) => {
-                const text = `${(percent * 100).toFixed(0)}%`;
-                return text;
-              }
-            }
-          }
-        }
-      },
-      {
-        // 面积图扩展。attribues 扩展 explaination （注意下，屏幕小的时候，布局自动变换为 上下）
-        attributes: {
-          // 抽取一个组件 Exlaination （组件注释下使用）
-          explaination: {
-            title: '专家分析',
-            details: [
-              {
-                icon: 'xx',
-                title: 'xx',
-                description: 'xx'
-              }
-            ]
-          }
-        }
-      },
-      {
-        tag: 'column',
-        id: 'f',
-        attributes: {
-          title: '业务域报警发布',
-          data:
-            'https://gw.alipayobjects.com/os/antfincdn/PC3daFYjNw/column-data.json',
-          isStack: true,
-          xField: 'city',
-          yField: 'value',
-          seriesField: 'type',
-          legend: {
-            position: 'top-left',
-          },
-        }
-      },
-      {
-        tag: 'line',
-        id: 'g',
-        attributes: {
-          title: '秒级总量监控',
-          data:
-            'https://gw.alipayobjects.com/os/antfincdn/PC3daFYjNw/column-data.json',
-          xField: 'city',
-          yField: 'value',
-          seriesField: 'type',
-          legend: false,
-          legendTable: {
-            show: true,
-            showColumnList: ['max', 'current']
-          }
-        }
-      }
-    ],
-  },
-};
